yahoonews_backend/requirements.txt
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.3.3
beautifulsoup4==4.8.2
blivet==3.1.0
Brlapi==0.6.7
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
configobj==5.0.6
coverage==4.5.1
cryptography==2.3
cupshelpers==1.0
dbus-python==1.2.4
decorator==4.2.1
Django==3.0.4
django-cors-headers==3.2.1
django-crontab==0.7.1
djangorestframework==3.11.0
gpg==1.10.0
html5lib==1.0.1
idna==2.7
initial-setup==0.3.62.1
isc==2.0
isort==4.3.21
langtable==0.0.38
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
mysqlclient==1.4.6
netifaces==0.10.6
ntplib==0.3.3
ordered-set==2.0.2
pciutils==2.3.6
perf==0.1
pid==2.1.1
ply==3.9
productmd==1.11
psycopg2==2.8.5
pwquality==1.4.0
pycairo==1.16.3
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pygobject==3.28.3
pykickstart==3.16.7
pylint==2.4.4
pymongo==3.10.1
pyOpenSSL==18.0.0
pyparted==3.11.0
PySocks==1.6.8
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.6
python-meh==0.47.2
pytz==2019.3
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
requests==2.20.0
requests-file==1.4.3
requests-ftp==0.3.1
rhnlib==2.8.6
rpm==4.14.2
schedutils==0.6
selinux==2.9
sepolicy==1.1
setools==4.2.2
setroubleshoot==1.1
simpleline==1.1
six==1.14.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.7
soupsieve==2.0
sqlparse==0.3.1
SSSDConfig==2.2.0
syspurpose==1.25.17
systemd-python==234
typed-ast==1.4.1
urllib3==1.24.3
webencodings==0.5.1
wrapt==1.11.2
